Much work! (and now we have a ChangeLog)
But, most of all, we have a quite full-working configuration dialog.
Now it is possible to modify Foopanel settings (seeing live the changes)
and to add and remove plugins from the configuration dialog.
There were some fixes and copyrights years bumps to plugins, too.

Also, SolidBlue-png theme is now called simply SolidBlue.

Still something needs to be finished, but this is a great jump towards
a Foopanel that can be used by everybody.